+++ packages/terra-dynamic-grid/tests/nightwatch/dynamic-grid-spec.js
@@ -1,40 +1,24 @@
-/* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
+/* eslint-disable no-unused-expressions */
 // eslint-disable-next-line import/no-extraneous-dependencies
-const screenshot = require('terra-toolkit').screenshot;
+const resizeTo = require('terra-toolkit/lib/nightwatch/responsive-helpers').resizeTo;
 
-module.exports = {
-  before: (browser, done) => {
-    browser.resizeWindow(browser.globals.width, browser.globals.height, done);
-  },
-
-  afterEach: (browser, done) => {
-    screenshot(browser, 'terra-dynamic-grid', done);
-  },
-
+module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], {
   'Displays a default dynamic-grid': (browser) => {
     browser
-      .url(`http://localhost:${browser.globals.webpackDevServerPort}/#/tests/dynamic-grid-tests/default`)
-      .assert.elementPresent('#defaultDynamicGrid');
+      .url(`${browser.launchUrl}/#/tests/dynamic-grid-tests/default`)
+      .expect.element('#defaultDynamicGrid').to.be.present;
 
-    /* TODO: Uncomment when we switch to headless chrome. PhantomJS doesn't support grids.
     browser.expect.element('#defaultDynamicGrid').to.have.css('display').which.matches(/^(-ms-)?grid$/);
     browser.expect.element('#defaultDynamicGrid').to.have.css('grid-gap');
     browser.expect.element('#defaultDynamicGrid').to.have.css('grid-template-columns');
     browser.expect.element('#defaultDynamicGrid').to.have.css('grid-template-rows');
-    */
 
-    /* TODO: Uncomment when we switch to headless chrome. PhantomJS doesn't support grids.
-    browser.expect.element('#region1').to.have.attribute('data-region-name').equals('r1');
     browser.expect.element('#region1').to.have.css('grid-row-start');
     browser.expect.element('#region1').to.have.css('grid-column-start');
-    */
     browser.expect.element('#region1').text.to.equal('R1');
 
-    /* TODO: Uncomment when we switch to headless chrome. PhantomJS doesn't support grids.
-    browser.expect.element('#region2').to.have.attribute('data-region-name').equals('r2');
     browser.expect.element('#region2').to.have.css('grid-row-start');
     browser.expect.element('#region2').to.have.css('grid-column-start');
-    */
     browser.expect.element('#region2').text.to.equal('R2');
   },
-};
+});