+++ packages/terra-embedded-content-consumer/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 -----------------
+### Changed
+* Update nightwatch tests and test scripts.
 
 1.0.0 - (August 31, 2017)
 ----------------